+++ src/senaite/core/browser/static/js/bika.lims.analysisrequest.add.js
@@ -273,6 +273,10 @@
        * Highlight a line number in the paste panel
        */
       this.highlight_paster_line = this.highlight_paster_line.bind(this);
+      /**
+       * Flush services search filters
+       */
+      this.flush_search_terms = this.flush_search_terms.bind(this);
       //#####################
       /* EVENT HANDLERS */
       //#####################
@@ -339,9 +343,21 @@
        */
       this.on_service_listing_header_click = this.on_service_listing_header_click.bind(this);
       /**
+       * Event handler for the services filter box
+       *
+       * Searches services their name match with the given term and make them
+       * visible. Categories without matches and without any pre-selected service
+       * are hidden. If no term, the visibility of categories and services is
+       * restored.
+       *
+       * @param event {Object} The event object
+       */
+      this.on_services_filter_change = this.on_services_filter_change.bind(this);
+      /**
        * Event handler for analysis service category rows.
        *
-       * Toggles the visibility of all services within this category.
+       * Toggles the visibility of all services within this category, but only if
+       * the search filter mode for this category is not enabled.
        * NOTE: Selected services always stay visible.
        *
        * @param event {Object} The event object
@@ -413,6 +429,8 @@
       this.deselected_uids = {};
       // Remove the '.blurrable' class to avoid inline field validation
       $(".blurrable").removeClass("blurrable");
+      // manually flush service search terms
+      this.flush_search_terms();
       // bind the event handler to the elements
       this.bind_eventhandler();
       // N.B.: The new AR Add form handles File fields like this:
@@ -593,6 +611,8 @@
       console.debug("AnalysisRequestAdd::bind_eventhandler");
       // Categories header clicked
       $("body").on("click", ".service-listing-header", this.on_service_listing_header_click);
+      // Categories filter search input changed
+      $("body").on("keyup", "input.services-filter", this.on_services_filter_change);
       // Category toggle button clicked
       $("body").on("click", "tr.category", this.on_service_category_click);
       // Composite Checkbox clicked
@@ -1505,6 +1525,80 @@
       }
     }
 
+    flush_search_terms() {
+      var el;
+      el = $("tr.service-listing-header input.services-filter");
+      return $(el).val("");
+    }
+
+    /**
+     * Returns the term to search through services from categories that belong
+     * to the given point of capture (poc)
+     *
+     * @param poc {String} The point of capture
+     * @returns {String} The search term
+     */
+    get_search_term(poc) {
+      var el, term;
+      el = $(`tr.${poc}.service-listing-header input.services-filter`);
+      // strip leading and trailing whitespaces
+      term = $(el).val();
+      return term.replace(/^\s+|\s+$/g, "");
+    }
+
+    /**
+     * Returns whether the search of services by term for the given point of
+     * capture is active
+     *
+     * @param poc {String} The point of capture
+     * @returns {Boolean} Whether the search by term is active or not
+     */
+    is_search_active(poc) {
+      var term;
+      term = this.get_search_term(poc);
+      if (term) {
+        return true;
+      }
+      return false;
+    }
+
+    /**
+     * Returns the list of service uids from categories that belong to the given
+     * point of capture and their human name match with the term passed-in.
+     * Returns an empty list if empty term
+     *
+     * @param poc {String} The point of capture
+     * @param term {String} The search term
+     * @returns {List} The list of service uids
+     */
+    search_services(poc, term) {
+      var matches, services;
+      matches = [];
+      // assume no matches if term is empty
+      if (!term) {
+        return matches;
+      }
+      // transform the term to lower case
+      term = term.toLowerCase();
+      // iterate through all registered services to find matches
+      services = $(`tr.${poc}.service`);
+      $.each(services, function(num, service) {
+        var $service, category_id, name, name_el, service_uid;
+        // get the service basic info
+        $service = $(service);
+        category_id = $service.data("category");
+        service_uid = $service.data("uid");
+        // get the service human name
+        name_el = $("div.service-title", $service);
+        name = name_el.html().toLowerCase();
+        // hide service if no match found and not checked for any sample
+        if (name.indexOf(term) !== -1) {
+          return matches.push(service_uid);
+        }
+      });
+      return matches;
+    }
+
     on_sample_nav(event) {
       var $el, el, idx, primary, target;
       el = event.currentTarget;
@@ -1770,12 +1864,82 @@
     }
 
     on_service_listing_header_click(event) {
-      var $el, poc, toggle, visible;
+      var $el, poc, services, toggle, visible;
       $el = $(event.currentTarget);
       poc = $el.data("poc");
+      // keep categories with selected services visible if search by term active
+      if (this.is_search_active(poc)) {
+        services = $(`tr.${poc}.service.visible`);
+        $.each(services, function(num, service) {
+          var $service, category;
+          $service = $(service);
+          category = $service.data("category");
+          return $(`tr.${poc}.${category}.category`).addClass("visible");
+        });
+        return;
+      }
+      // toggle visibility
       visible = $el.hasClass("visible");
       toggle = !visible;
       return this.toggle_poc_categories(poc, toggle);
+    }
+
+    on_services_filter_change(event) {
+      var $el, categories, expanded_categories, poc, services, term, uids;
+      $el = $(event.currentTarget);
+      poc = $el.closest("tr").data("poc");
+      // keep track of the categories to expand and keep visible
+      expanded_categories = new Set([]);
+      // get the term so search by
+      term = this.get_search_term(poc);
+      // do the search
+      uids = this.search_services(poc, term);
+      // show services that match or with ticked checkboxes
+      services = $(`tr.${poc}.service`);
+      $.each(services, function(num, service) {
+        var $service, category, checked, uid;
+        $service = $(service);
+        category = $service.data("category");
+        uid = $service.data("uid");
+        if (indexOf.call(uids, uid) >= 0) {
+          $service.addClass("visible");
+          expanded_categories.add(category);
+          return;
+        }
+        // show if ticked checkboxes
+        checked = $("input[type=checkbox]:checked", $service);
+        if (checked.length > 0) {
+          $service.addClass("visible");
+          expanded_categories.add(category);
+          return;
+        }
+        // hide the service
+        return $service.removeClass("visible");
+      });
+      // expand categories with visible services
+      categories = $(`tr.${poc}.category`);
+      return $.each(categories, function(num, category) {
+        var $btn, $category, category_id, expanded;
+        // toggle the visibility of the category
+        $category = $(category);
+        $btn = $(".service-category-toggle", $category);
+        category_id = $category.data("category");
+        // expand the category if at least on service visible
+        expanded = expanded_categories.has(category_id);
+        if (expanded) {
+          $category.addClass("expanded");
+          $btn.text("-");
+        } else {
+          $category.removeClass("expanded");
+          $btn.text("+");
+        }
+        // make category visible if expanded or empty search term
+        if (expanded || !term) {
+          return $category.addClass("visible");
+        } else {
+          return $category.removeClass("visible");
+        }
+      });
     }
 
     on_service_category_click(event) {
@@ -1783,6 +1947,10 @@
       event.preventDefault();
       $el = $(event.currentTarget);
       poc = $el.attr("poc");
+      // do nothing if search by term is active
+      if (this.is_search_active(poc)) {
+        return;
+      }
       $btn = $(".service-category-toggle", $el);
       expanded = $el.hasClass("expanded");
       category = $el.data("category");