+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - #2698 Custom contact widget in user profile
+- #2699 Add search filter by term for services in add sample form
 - #2692 Refactor Sticker Functionality
 - #2696 Fix error when removing a Worksheet from inside its view
 - #2690 Support date and datetime on result entry